+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 + [x] Events: fix `fire`. Now main is constructor `CustonEvent`, for old - `createEvent`
 + [x] Manipulation: add method `wrap`, `wrapAll`, `wrapInner`. The method puts elements inside the wrapper and return `wrapper(s)`. 
 + [x] Init: change `throw Error` to `console.warn` when selector is `#` or `.`
++ [x] setImmediate: add support for `process` and `web workers` 
 
 ### 1.0.6
 + [x] Init: fix minor bug for creating elements in context